# docs/swagger.yaml
definitions:
  admin_dto.Album:
    properties:
      created_at:
        type: integer
      id:
        type: string
      participant_count:
        type: integer
      title:
        type: string
      total_cards:
        type: integer
    type: object
  admin_dto.CreateCardRequest:
    properties:
      album_id:
        type: string
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      number:
        type: string
    type: object
  admin_dto.Message:
    properties:
      message:
        type: string
    type: object
...
  admin_dto.UpdateAlbumRequest:
    properties:
      title:
        type: string
    type: object
  admin_dto.UpdateCardRequest:
    properties:
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      number:
        type: string
    type: object
...
  album_dto.CreateAlbumCardRequest:
    properties:
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      number:
        type: string
    type: object
  album_dto.CreateAlbumRequest:
    properties:
      cards:
        items:
          $ref: '#/definitions/album_dto.CreateAlbumCardRequest'
        type: array
      title:
        type: string
    type: object
  album_dto.RegisterCardRequest:
    properties:
      album_id:
        type: string
      card_id:
        type: string
      contact_id:
        type: string
      secret:
        type: string
    type: object
  album_model.Album:
    properties:
      cards:
        items:
          $ref: '#/definitions/card_model.Card'
        type: array
      created_at:
        type: integer
      id:
        type: string
      title:
        type: string
      total_cards:
        type: integer
    type: object
  album_participant_dto.CreateAlbumParticipantRequest:
    properties:
      album_id:
        type: string
    type: object
  album_participant_dto.ShareAssignedCardResponse:
    properties:
      album_id:
        type: string
      card_id:
        type: string
      contact_id:
        type: string
      secret:
        type: string
    type: object
  album_participant_model.AlbumParticipant:
    properties:
      album_id:
        type: string
      assigned_card_id:
        type: string
      joined_at:
        type: integer
      user_id:
        type: string
    type: object
  card_dto.CreateCardRequest:
    properties:
      album_id:
        type: string
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      number:
        type: string
    type: object
  card_model.Card:
    properties:
      album_id:
        type: string
      created_at:
        type: integer
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      number:
        type: string
      updated_at:
        type: integer
    type: object
...
  sqlc.Album:
    properties:
      createdAt:
        format: int64
        type: integer
      id:
        type: string
      title:
        type: string
      totalCards:
        format: int64
        type: integer
    type: object
...
  user_model.User:
    properties:
      created_at:
        type: integer
      description:
        type: string
      email:
        type: string
      github:
        type: string
      id:
        type: string
      is_admin:
        type: integer
      linkedin:
        type: string
      name:
        type: string
      picture:
        type: string
      update_at:
        type: integer
      web:
        type: string
    type: object
info:
  contact: {}
  description: MVP para intercambio de tarjetas coleccionables.
  title: Trading Card Album API
  version: "1.0"
paths:
  /api/v1/admin/albums:
    get:
      consumes:
      - application/json
      description: List all albums with stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin_dto.Album'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create an album with its cards
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/album_dto.CreateAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/album_model.Album'
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/albums/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_dto.Message'
      security:
      - BearerAuth: []
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update album title
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin_dto.UpdateAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_dto.Album'
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/albums/{id}/cards:
    get:
      consumes:
      - application/json
      description: Get cards of an album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/card_model.Card'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/cards:
    get:
      consumes:
      - application/json
      description: List all cards
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/card_model.Card'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a card in an album
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin_dto.CreateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/card_model.Card'
      security:
      - BearerAuth: []
      tags:
      - Admin
  /api/v1/admin/cards/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a card
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_dto.Message'
      security:
      - BearerAuth: []
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update a card
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin_dto.UpdateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/card_model.Card'
      security:
      - BearerAuth: []
      tags:
      - Admin
...
  /api/v1/album:
    get:
      consumes:
      - application/json
      description: Get albums by user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sqlc.Album'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Album
    post:
      consumes:
      - application/json
      description: Create album
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/album_dto.CreateAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/album_model.Album'
      security:
      - BearerAuth: []
      tags:
      - Album
  /api/v1/album/{id}:
    get:
      consumes:
      - application/json
      description: Get album by id
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/album_model.Album'
      security:
      - BearerAuth: []
      tags:
      - Album
  /api/v1/album/{id}/assigned_card:
    get:
      consumes:
      - application/json
      description: Assigned card
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/card_model.Card'
      security:
      - BearerAuth: []
      tags:
      - Album
  /api/v1/album/{id}/card:
    get:
      consumes:
      - application/json
      description: Card Pool
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/card_model.Card'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Album
  /api/v1/album/{id}/join_qr:
    get:
      consumes:
      - application/json
      description: Get join QR of an album (encode del código de invitación)
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      tags:
      - Album
  /api/v1/album/{id}/qr_events:
    get:
      consumes:
      - application/json
      description: 'SSE: avisa cuando el QR compartido de este álbum fue escaneado'
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      tags:
      - Album
  /api/v1/album/{id}/share_assigned_card:
    get:
      consumes:
      - application/json
      description: Share assigned card
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: string
      - description: qr view
        in: query
        name: qr
        type: string
      produces:
      - application/json
      - image/png
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/album_participant_dto.ShareAssignedCardResponse'
      security:
      - BearerAuth: []
      tags:
      - Album
  /api/v1/album/new_card:
    post:
      consumes:
      - application/json
      description: Register new card
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/album_dto.RegisterCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/card_model.Card'
      security:
      - BearerAuth: []
      tags:
      - Album
  /api/v1/album_participant:
    post:
      consumes:
      - application/json
      description: Create Album participant
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/album_participant_dto.CreateAlbumParticipantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/album_participant_model.AlbumParticipant'
      security:
      - BearerAuth: []
      tags:
      - Album Participant
  /api/v1/auth/me:
    get:
      consumes:
      - application/json
      description: Get current session user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_model.User'
      security:
      - BearerAuth: []
      tags:
      - Auth
  /api/v1/card:
    post:
      consumes:
      - application/json
      description: Create card
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/card_dto.CreateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/card_model.Card'
      security:
      - BearerAuth: []
      tags:
      - Card
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"